Include a link to pgloader
This script has a very good Google rank for migrating from MySQL to PostgreSQL and seems to be quite dedicated to a specific environment and migration project. Pgloader (https://github.com/dimitri/pgloader) solves the generic problem and work directly from a MySQL connection, you can run it in your CI environment until you're ready to switch in production. As pgloader author, it seems to me that linking the project from your README would be valuable for the both of us, allowing you to focus on your context and helping the general case being better covered.